refactor(artifact-provenance): extract read model
